+++ HISTORY.rst
@@ -8,6 +8,7 @@ ______________________
 
 - If not explicitly declared, `Part` objects will load the default footprint from their symbol definition.
 - Added `empty_footprint_handler()` for parts without footprints that logs errors by default but can be overriden by the user.
+- Gitlab KiCad V6 symbol libraries can be used by placing `https://gitlab.com/kicad/libraries/kicad-symbols/-/raw/master` in `lib_search_paths[KICAD]`.
 
 
 1.1.0 (2021-10-22)